bench: add request header benchmarks
