+++ packages/node-fetch-server/bench/runner.sh
@@ -32,12 +32,16 @@ echo $(node -e 'console.log(`Date: ${new Date().toLocaleString()}`)')
 NODE_VERSION=$(node -e 'console.log(process.version.slice(1))')
 run_benchmark "node:http@$NODE_VERSION" \
   "node ./servers/node-http.ts"
+run_benchmark "node:http-request-headers@$NODE_VERSION" \
+  "node ./servers/node-http-request-headers.ts"
 
 NODE_FETCH_SERVER_VERSION=$(node -e 'console.log(require("../package.json").version)')
 run_benchmark "node-fetch-server@$NODE_FETCH_SERVER_VERSION" \
   "node ./servers/node-fetch-server.ts"
 run_benchmark "node-fetch-server-request@$NODE_FETCH_SERVER_VERSION" \
   "node ./servers/node-fetch-server-request.ts"
+run_benchmark "node-fetch-server-request-headers@$NODE_FETCH_SERVER_VERSION" \
+  "node ./servers/node-fetch-server-request-headers.ts"
 
 EXPRESS_VERSION=$(node -e 'console.log(require("express/package.json").version)')
 run_benchmark "express@$EXPRESS_VERSION" \
